+++ src/strategyCalculator/Strategy.java
@@ -1,7 +1,7 @@
 package strategyCalculator;
 
 public interface Strategy {
-	public int DoOpertion(int a, int b);
+	public float DoOpertion(float a, float b);
 }
